@remotion/studio: Fix marquee selection from empty timeline area
