+++ packages/studio/src/components/Timeline/TimelineScrollable.tsx
@@ -1,7 +1,10 @@
 import React, {useMemo} from 'react';
 import {HORIZONTAL_SCROLLBAR_CLASSNAME} from '../Menu/is-menu-item';
 import {scrollableRef} from './timeline-refs';
-import {TIMELINE_BACKGROUND} from './TimelineSelection';
+import {
+	TIMELINE_BACKGROUND,
+	useTimelineMarqueeSelection,
+} from './TimelineSelection';
 
 const outer: React.CSSProperties = {
 	width: '100%',
@@ -12,9 +15,19 @@ const outer: React.CSSProperties = {
 	backgroundColor: TIMELINE_BACKGROUND,
 };
 
+const marqueeStyle: React.CSSProperties = {
+	backgroundColor: 'rgba(70, 130, 255, 0.16)',
+	border: '1px solid rgba(70, 130, 255, 0.75)',
+	boxSizing: 'border-box',
+	pointerEvents: 'none',
+	position: 'fixed',
+	zIndex: 10,
+};
+
 export const TimelineScrollable: React.FC<{
 	readonly children: React.ReactNode;
 }> = ({children}) => {
+	const {marqueeRect, onPointerDownCapture} = useTimelineMarqueeSelection();
 	const containerStyle: React.CSSProperties = useMemo(() => {
 		return {
 			width: '100%',
@@ -27,8 +40,20 @@ export const TimelineScrollable: React.FC<{
 			ref={scrollableRef}
 			style={outer}
 			className={HORIZONTAL_SCROLLBAR_CLASSNAME}
+			onPointerDownCapture={onPointerDownCapture}
 		>
 			<div style={containerStyle}>{children}</div>
+			{marqueeRect === null ? null : (
+				<div
+					style={{
+						...marqueeStyle,
+						height: marqueeRect.bottom - marqueeRect.top,
+						left: marqueeRect.left,
+						top: marqueeRect.top,
+						width: marqueeRect.right - marqueeRect.left,
+					}}
+				/>
+			)}
 		</div>
 	);
 };
